requirements.txt
Flask 2.0.1
Jinja2==2.8
vonage==2.5.5
requests==2.20.0
python-dotenv==0.19.0